Fixes fatal error Class 'Mandrill_Error' not found
In case an invalid error response comes back, the error object is casted to a `Postman_Mandrill_Error`. However, if no name is set, or the status is not 'error' it falls back to a dynamic error name. The class referenced there (`Mandrill_Error`) does not exist, and should be `Postman_Mandrill_Error`